Refactor Head Parsing to ease porting to main branch
