+++ docs/modules/site-integration/pages/exposed-metadata.adoc
@@ -1,16 +1,17 @@
 = Exposed metadata
 :asciidoctor-docs-url: https://docs.asciidoctor.org/asciidoc/latest
+:maven-site-plugin-docs-url: https://maven.apache.org/plugins/maven-site-plugin
 
-The Asciidoctor Maven Site integration integrates with Doxia to expose some of its information.
+The Asciidoctor Maven Site integration collaborates with Doxia to expose some of its information.
 
 == Document Header metadata
 
-The following information from the {asciidoctor-docs-url}/document/header/[header] is extracted:
+The following elements from the {asciidoctor-docs-url}/document/header/[header] are integrated:
 
-document title:: used to inform the breadcrumb line when these are enabled.
+document title:: used to inform the {maven-site-plugin-docs-url}/examples/sitedescriptor.html#Breadcrumbs[breadcrumb] line when these are enabled.
 
 author(s):: full representation (full name and email) will be present as HTML `<meta name="author" ... >` tags inside the HTML `<head>`.
-In case of multiple authors, each one will appear in a distinct meta element.
+In case of multiple authors, each one will appear in a distinct `meta` element.
 
-revision date:: the header revision date value wll be presented as-is in a `<meta name="date" ... >` element.
+revision date:: the header revision date value will be presented as-is in a `<meta name="date" ... >` element.
 Alternatively, if not set, the generated value of `docdatetime` will be used.
